multi-crate command
