+++ Cargo.toml
@@ -1,3 +1,10 @@
+[workspace]
+members = [
+    ".",
+    "controller",
+    "load",
+]
+
 [package]
 name = "ortiofay"
 version = "0.1.0"
@@ -6,13 +13,8 @@ edition = "2018"
 publish = false
 
 [dependencies]
-futures = "0.3"
-indexmap = "1"
-kube = { git = "https://github.com/clux/kube-rs", default-features = false, features = ["derive", "rustls-tls"] }
-k8s-openapi = { version = "0.10.0", default-features = false, features = ["v1_19"] }
-serde = { version = "1.0.111", features = ["derive"] }
-serde_json = "1.0.57"
-serde_yaml = "0.8.13"
-tokio = { version = "0.2", features = ["macros", "sync"] }
-tracing = "0.1"
+ortiofay-controller = { path = "./controller" }
+ortiofay-load = { path = "./load" }
+structopt = "0.3"
+tokio = { version = "0.2", features = ["macros"] }
 tracing-subscriber = "0.2"
