+++ controller/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "ortiofay-controller"
+version = "0.1.0"
+authors = ["Oliver Gould <ver@buoyant.io>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+futures = "0.3"
+indexmap = "1"
+kube = { git = "https://github.com/clux/kube-rs", default-features = false, features = ["derive", "rustls-tls"] }
+k8s-openapi = { version = "0.10.0", default-features = false, features = ["v1_19"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+serde_yaml = "0.8"
+structopt = "0.3"
+tokio = { version = "0.2", features = ["macros", "sync"] }
+tracing = "0.1"
