+++ load/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ortiofay-load"
+version = "0.1.0"
+authors = ["Oliver Gould <ver@buoyant.io>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+futures = "0.3"
+http = "0.2"
+hyper = "0.13.3"
+structopt = "0.3"
+tokio = { version = "0.2", features = ["macros"] }
+tracing = "0.1"
